split functions to separate aggregates
